+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-v1.0, since 2016-10-12
+v1.0, since 2016-10-06
 ----------------------
 
+- Added `--single-build-dir` command-line option
 - Added `GIT_TAG_OVERRIDE` option for `add_pkg`/`def_pkg` commands
 - Added using `<exe-dir>/default-cmakex-presets.yaml` if CMAKEX_PRESET_FILE
   is not set
 - Added `--deps-source`, `--deps-build` and `--deps-install` command-line
   options
+- Added manifest generation with `--manifest`
+- Added per-config build dir feature and made it default